pyproject.toml
[project]
name = "ease-of-use-scripts"
version = "0.0.7"
description = ""
requires-python = ">=3.8,<4.0.0"
dependencies = [
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
cqc = "ease_of_use_scripts.cqc:main"
setup = "ease_of_use_scripts.setup_new_environment:main"
run_pytest = "ease_of_use_scripts.run_pytest:main"
